[refactor] Move all operands into a single file
